* javadoc: added overview.html with img of architecture.jpg   * moved new targets (javadoc, updatesite) to main ant-build-file   * remove redundant build-files   * fix an error I introduced to test exceptions ...
git-svn-id: https://vimplugin.svn.sourceforge.net/svnroot/vimplugin/trunk@221 e026b6f4-af2c-0410-b418-94eebdd9c5e6

+++ plugin/src/org/vimplugin/editors/AbstractVimEditor.java
@@ -209,11 +209,8 @@ public class AbstractVimEditor extends TextEditor {
 		} else {
 			f = c.getField(AbstractVimEditor.win32WID);
 		}
-		//Correct:
-		//wid = f.getLong(parent);
-		//BUG:
-		wid = f.getInt(parent);
-		
+
+		wid = f.getLong(parent);
 		
 		int h = parent.getClientArea().height;
 		int w = parent.getClientArea().width;
